Moved .config into a cog, and made it add to a config
.setconfig was made and did what .config did previously

.config and .setconfig where made useable to allowRoleList for only config values in staffConfig
